requirements.txt
APScheduler==3.7.0
Flask==1.1.4
Flask-APScheduler==1.12.2
Flask-Caching==1.10.1
flask-restx==0.4.0
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
python-dotenv==0.18.0
gevent==21.1.2
mysql-connector-python==8.0.26
python-dateutil==2.8.1
pytest==6.2.4
pytest-mock==3.6.1
SQLAlchemy==1.4.22
redis==3.5.3
fortinet-common==0.2